La classe Database fonctionne. Elle gere la connexion en BD
